+++ scripts/custom_nodes/selection_container.gd
@@ -2,5 +2,5 @@ extends VBoxContainer
 class_name SelectionContainer
 
 func get_buttons() -> Array:
-	Utils.temp_out = []
+	
 	return Utils.get_children_from_type(self, Button)